\name{pamm-package}
\alias{pamm-package}
\alias{pamm}
\docType{package}
\title{
Power analysis for random effects in mixed models
}
\description{
Description: Simulation functions to assess or explore the power of a dataset to estimates significant random effects (intercept or slope) in a mixed model. The functions are based on the "lme4" package.
}
\details{
\tabular{ll}{
Package: \tab pamm\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2009-06-02\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Julien Martin

Maintainer:Julien Martin <julien.martin2@usherbrooke.ca>
}
\references{
Martin, Nussey, Wilson and Reale ???? Power analysis for random
effects. Ecology Letters, ????
}

\keyword{ package }
\seealso{
}
\examples{
}
