\name{plot.SSF}
\alias{plot.SSF}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ GRAPHIC OUTPUT OF THE SSF FUNCTION }
\description{
  provide graphic interpretation of the simulation results
}
\usage{
\method{plot}{SSF}(x,\dots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ an SSF object }
  \item{\dots}{ potentially further arguments to pass to methods }
}
\details{ }
\value{ }
\references{ }
\author{ Julien Martin }
\note{ }
\seealso{ \code{\link{SSF}},\code{\link{plot.PAMM}},\code{\link{plot.EAMM}} }
\examples{
## Not run:
#  ours<- SSF(10,200,10,c(0.4,0.1,0.6,0))
#  plot(ours)
## End(Not run)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ misc }

